+++ src/projectCoupon/facad/CustomerFacad.java
@@ -15,8 +15,8 @@ import projectCoupon.dbdao.Customer_CouponDBDAO;
 import projectCoupon.exception.CouponException;
 import projectCoupon.exception.CreateException;
 import projectCoupon.exception.CustomerException;
+import projectCoupon.exception.RemoveException;
 import projectCoupon.exception.UpdateException;
-import projectCoupon.utils.ClientType;
 import projectCoupon.utils.Utile;
 
 public class CustomerFacad implements CouponClientFacade {
@@ -33,24 +33,35 @@ public class CustomerFacad implements CouponClientFacade {
 		customer_CouponDAO = new Customer_CouponDBDAO();
 	}
 
-
-
-	public void insertCustomer(Customer customer) throws Exception {
+	public void insertCustomer(Customer customer) throws CouponException, CreateException {
+		if (custId == 0) {
+			throw new CouponException("the operation was canceled due to not being loged in");
+		}
 		custDAO.insertCustomer(customer);
 	}
 
-	public void removeCustomer(Customer customer) throws Exception {
+	public void removeCustomer(Customer customer) throws CouponException, RemoveException {
+		if (custId == 0) {
+			throw new CouponException("the operation was canceled due to not being loged in");
+		}
 		customer_CouponDAO.removeCustomer_Coupon(customer);
 		custDAO.removeCustomer(customer);
 	}
 
-	public void updateCustomer(Customer customer, String newName, String newpassword) throws Exception {
+	public void updateCustomer(Customer customer, String newName, String newpassword)
+			throws CouponException, UpdateException {
+		if (custId == 0) {
+			throw new CouponException("the operation was canceled due to not being loged in");
+		}
 		customer.setCustomerName(newName);
 		customer.setPassword(newpassword);
 		custDAO.updateCustomer(customer);
 	}
 
-	public void purchaseCoupon(long coupId) throws CouponException, CreateException, UpdateException{
+	public void purchaseCoupon(long coupId) throws CouponException, CreateException, UpdateException { 
+		if(custId == 0) {
+			throw new CouponException("the operation was canceled due to not being loged in");
+		}
 		Coupon coupon = new Coupon();
 		coupon = couponDAO.getCoupon(coupId);
 		if (coupon != null) {
@@ -71,31 +82,45 @@ public class CustomerFacad implements CouponClientFacade {
 			}
 		} else {
 			throw new CouponException("Coupon Information Not Exist! Purchase Failed!");
+
 		}
 	}
 
-	public Coupon getCoupon(long coupId) throws Exception {
+	public Coupon getCoupon(long coupId) throws CouponException, CreateException {
+		if(custId == 0) {
+			throw new CouponException("the operation was canceled due to not being loged in");
+		}
 		return couponDAO.getCoupon(coupId);
+
 	}
 
 	public List<Coupon> getAllCouponsByType(couponType couponType) throws CouponException {
+		if(custId == 0) {
+			throw new CouponException("the operation was canceled due to not being loged in");
+		}
 		return couponDAO.getAllCouponsByType(couponType);
 	}
 
 	public List<Coupon> getAllPurchasedCoupons() throws CouponException, CreateException {
+		if(custId == 0) {
+			throw new CouponException("the operation was canceled due to not being loged in");
+		}
 		List<Long> customersCoupons = customer_CouponDAO.getCouponsByCustomerId(custId);
 		List<Coupon> purchasedCoupons = new ArrayList<Coupon>();
-		for(Long id:customersCoupons) {
+		for (Long id : customersCoupons) {
 			purchasedCoupons.add(couponDAO.getCoupon(id));
 		}
 		return purchasedCoupons;
 	}
 
 	public List<Coupon> getAllPurchasedCouponsByType(couponType type) throws CouponException, CreateException {
+		if(custId == 0) {
+			throw new CouponException("the operation was canceled due to not being loged in");
+		}
 		List<Coupon> coupons = getAllPurchasedCoupons();
 		List<Coupon> couponByType = new ArrayList<Coupon>();
 		for (Coupon coupon : coupons) {
-			if (coupon.getType().name().equals(type.name())){
+			if (coupon.getType().name().equals(type.name())) {
 				couponByType.add(coupon);
 			}
 		}
@@ -110,50 +135,64 @@ public class CustomerFacad implements CouponClientFacade {
 	 *            double
 	 * @return Coupon collection
 	 * @throws CouponException
-	 * @throws CreateException 
+	 * @throws CreateException
 	 */
 	public List<Coupon> getAllPurchasedCouponsByPrice(long price) throws CouponException, CreateException {
+		if(custId == 0) {
+			throw new CouponException("the operation was canceled due to not being loged in");
+		}
 		List<Coupon> coupons = getAllPurchasedCoupons();
 		List<Coupon> couponByType = new ArrayList<Coupon>();
 		for (Coupon coupon : coupons) {
-			if (coupon.getPrice() == price){
+			if (coupon.getPrice() == price) {
 				couponByType.add(coupon);
 			}
 		}
 		return couponByType;
 	}
 
-	public long getCustomerId() {
+	public long getCustomerId() throws CouponException {
+		if(custId == 0) {
+			throw new CouponException("the operation was canceled due to not being loged in");
+		}
 		return custId;
 	}
 
 	/**
 	 * @return Customer
+	 * @throws CouponException 
 	 */
-	public Customer getCustomerInstance() {
+	public Customer getCustomerInstance() throws CouponException {
+		if(custId == 0) {
+			throw new CouponException("the operation was canceled due to not being loged in");
+		}
 		return customer;
 	}
 
-	public Customer getCustomer() throws CustomerException{
-			System.out.println(custDAO.getCustomer(this.customer.getCustomerId()));
-			return custDAO.getCustomer(this.customer.getCustomerId());
+	public Customer getCustomer() throws CustomerException, CouponException {
+		if(custId == 0) {
+			throw new CouponException("the operation was canceled due to not being loged in");
+		}
+		System.out.println(custDAO.getCustomer(this.customer.getCustomerId()));
+		return custDAO.getCustomer(this.customer.getCustomerId());
 	}
 
-	public java.util.List<Customer> getAllCustomer() throws Exception {
+	public java.util.List<Customer> getAllCustomer() throws CustomerException, CouponException{
+		if(custId == 0) {
+			throw new CouponException("the operation was canceled due to not being loged in");
+		}
 		return custDAO.getAllCustomer();
 	}
 
-	public static CouponClientFacade login(String name, String password) throws Exception {
+	public CouponClientFacade login(String name, String password) throws Exception {
 		Customer customer = new CustomerDBDAO().login(name, password);
 		if (customer != null) {
-			CustomerFacad customerFacad = new CustomerFacad();
-			customerFacad.custId = customer.getCustomerId();
-			customerFacad.customer = customer;
-			return customerFacad;
+			this.custId = customer.getCustomerId();
+			this.customer = customer;
+			return this;
 		} else {
 			return null;
 		}
-		
-		
+
 	}
 }
